requirements.txt
dotenv==0.9.9
pytest==8.3.5
python-dotenv==1.1.0
selenium==4.31.0
allure-pytest==2.14.2
pytest-xdist==3.6.1